Create pm_nova_mainscript
This is code that changes the original script and loops the code to log the data every second and save the output as a CSV file. This file gives very important insights into the data logged from PM Sensor, The original code was modified to make it run in a loop. The current code prints the data continuously as well as logs the information in CSV format.